Change: rename from UserAuthForm to LoginForm
